+++ plugin/skwirrel-pim-sync/skwirrel-pim-sync.php
@@ -4,10 +4,10 @@
  * Plugin Name: Skwirrel PIM sync for WooCommerce
  * Plugin URI: https://github.com/Skwirrel-B-V/skwirrel-pim-sync-for-woocommerce
  * Description: Sync plugin for Skwirrel PIM via Skwirrel JSON-RPC API to WooCommerce.
- * Version: 3.10.1
+ * Version: 3.10.2
  * Author: Skwirrel B.V.
  * Author URI: https://skwirrel.eu
- * Requires at least: 6.0
+ * Requires at least: 6.9
  * Requires PHP: 8.3
  * WC requires at least: 8.0
  * WC tested up to: 10.6
@@ -25,7 +25,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'SKWIRREL_WC_SYNC_VERSION', '3.10.1' );
+define( 'SKWIRREL_WC_SYNC_VERSION', '3.10.2' );
 define( 'SKWIRREL_WC_SYNC_PLUGIN_FILE', __FILE__ );
 define( 'SKWIRREL_WC_SYNC_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 define( 'SKWIRREL_WC_SYNC_PLUGIN_URL', plugin_dir_url( __FILE__ ) );